Update nodeserver enable state during start/stop
The nodeserver database table containes a field called enabled which
is set to true by default.  Make use of that field to track the
current user specified state of the nodeserver.

If the user stops the nodeserver, update the enabled field to false
If the user starts or restarts the nodeserver, update the enabled field
to true.

When starting Polyglot, use the nodeserver's enabled field to determine
if the nodeserver should be automatically started or not.  If enabled
is true, start the nodeserver, if it is false, don't start the nodeserver.

Signed-off-by: Bob Paauwe <bpaauwe@yahoo.com>