Merge/rebase local changes when starting nodeserver.
When starting a nodeserver, a git pull is done to make sure the
most recent version is being used.  However, without options, this
will break if the user has commited addional changes or has switched
to a new branch.

This situtation is most likely to occure when someone is doing
active development/debugging of an existing nodeserver

The simple solution is to use the --rebase option on git pull. This
will pull the most current version and then rebase any local changes
on top.

In most cases, this will do the right thing.  It should only abort
if there are conflicts between the local changes and the updated
code.  In that case, it is up to the user to resolve those conficts
before they'll be able to run the nodeserver.

Signed-off-by: Bob Paauwe <bpaauwe@yahoo.com>