Clean process id when stopping nodeserver
When the nodeserver is stopped, make sure we delete the process
id so that we can restart the nodeserver. Otherwise it is possible
to get stuck in a loop where it is impossible to start the nodeserver
because it thinks it is already running and stopping it doesn't do
anything because it's not really running.

Signed-off-by: Bob Paauwe <bpaauwe@yahoo.com>