Make store URL configurable
Add a text box to the Polyglot settings screen where the nodeserver
store URL can be entered. By default, it is set to the value that
was hard coded.

   https://pgcstore.isy/io/

When the store is queried, it appends v1/list?sort to the URL as that
is apparently the store API for getting a list of nodeservers so keep
this in mind if you are replacing the URL with your own.

Eventual goal would be to have this be a selection box that allows
only approved nodeserver stores to be queried.  Most likely, this
would include a production store and a test store for developers.

Signed-off-by: Bob Paauwe <bpaauwe@yahoo.com>