Add support for setting node server log level.
Add a drop down control to the log section of the dashboard UI
that allows the user to set a desired log level.  The level is
sent to the node server.

In addition, save the current level in the nodeserver's db so
that it may be persistent across restarts.

Initially, the log level is read from the server.json file if
present, if not, the default is Warning.

The node server may also send an updated log level and that will
be reflected in the UI and stored in the db.

The list of available log levels is hard coded in the UI.