Add a wait before sending config via stdin.
There can be a race condition when a nodeserver exits before we have
sent the configuration info via STDIN. We may attempt to send the info
after the nodeserver has exited but before we deleted the process id
to indicate it's no longer active.

Adding a wait here gives us a chance to delete the process id before
attempting to send the configuration info.
If the nodeserver exits on start before we send the configuration
infomation, we'll crash because the file handle is go