Add private node data database field.
Add a field to the node database table to hold node specific
private data.  We use a BLOB and allow it to contain any data
similar to how the custom database fields are implimented.

This data is populated when the node is first created.  It should
be considered to be immutable after that. This is similar to other
node data fields like node address.

The node server should set the private data before calling
addNode() to add the node to the database.

The general use case for this is to hold something like a hardware
device ID or device token that will allow the node server to map
the node back to a hardware device when restarting.