Client configurable log level list
The list of allowable log levels is stored in the nodeserver
database.  By default, the list matches the named levels in
the LOGGER.

Added API to allow the client to update this list and have
the updated list be shown in place of the default list.

This allows each node server to customize the list of
allowable log levels by either restricting it or adding
new node server specific levels.