ui: functions to access the UDI Portal API
portalPurchaseNodeServer()  Redirect to the Portal to purchase
a node server.  This allows the user to securely purchase a license
to install a node server. This is only applicable to non-free node
servers.

portalSyncTransactions() Query the UDI Portal system for a list
of purchase transactions for this Polisy device. This returns a
list of node server purchased and info about the purchase.  Used
to fill out the available node server table.

portalNodeServerInstalled() Notify the UDI Portal system if a
node server has been installed or uninstalled.