ui: Node server store listing updates for purchase options
Scan the various purchase options associated with a node server
to determine the price. It is possible for the node server to have
multiple purchase options but only the first option with a price
field is used.  All options will be shown when redirected to the
Portal to purchase the node server.

Get the list of purchase transactions.  This will be used to determine
if the we should attempt to install the node server (valid license
exists) or re-direct to the Portal to purchase a license when the
user clicks the "Install" button.

Redirect to the Portal to purchase a license if none exists for the
node server.

When a node server is installed, notify the Portal that it was installed.