Node server subscription exiration support
If a node server was purchased for a specific subscription period,
add the expriration date to the database and check running node servers
to make sure the subscription hasn't expired.

Provide a warning a few days in advance by posting a notice to the
dashboard.

If a node server subscription is expired, stop the node server.

This checks the expiration every minute.

Note, this also applies to a trial subscription.